(pd) factor out "steplist" Redux code
